Add Divisible protocol, create stub Line struct, fix integer conversion in Color and implement vector scaling

+++ Sources/D2/util/graphics/Color.swift
@@ -4,9 +4,9 @@ struct Color {
 	let blue: UInt8
 	let alpha: UInt8
 	
-	var rgb: UInt32 { return (red << 16) | (green << 8) | blue }
-	var rgba: UInt32 { return (red << 24) | (green << 16) | (blue << 8) | alpha }
-	var argb: UInt32 { return (alpha << 24) | (red << 16) | (green << 8) | blue }
+	var rgb: UInt32 { return UInt32((red << 16) | (green << 8) | blue) }
+	var rgba: UInt32 { return UInt32((red << 24) | (green << 16) | (blue << 8) | alpha) }
+	var argb: UInt32 { return UInt32((alpha << 24) | (red << 16) | (green << 8) | blue) }
 	
 	init(red: UInt8, green: UInt8, blue: UInt8, alpha: UInt8 = 255) {
 		self.red = red
